Move normalizations to their own file to avoid circular imports
