+++ desc/objectives/_equilibrium.py
@@ -15,8 +15,8 @@ from desc.grid import ConcentricGrid, QuadratureGrid
 from desc.transform import Transform
 from desc.utils import Timer
 
+from .normalization import compute_scaling_factors
 from .objective_funs import _Objective
-from .utils import compute_scaling_factors
 
 
 class ForceBalance(_Objective):
